refactor: rename pydantic stub
